Orchestrator/requirements.txt
# requirements.txt
psycopg2-binary==2.9.9 
requests==2.31.0
schedule==1.2.1
matplotlib==3.8.2

